+++ src/code42cli/cmds/departing_employee.py
@@ -4,6 +4,7 @@ from py42.exceptions import Py42NotFoundError
 from code42cli.bulk import generate_template_cmd_factory
 from code42cli.bulk import run_bulk_process
 from code42cli.click_ext.groups import OrderedGroup
+from code42cli.cmds.detectionlists import HEADER_KEYS_MAP
 from code42cli.cmds.detectionlists import update_user
 from code42cli.cmds.detectionlists.options import cloud_alias_option
 from code42cli.cmds.detectionlists.options import notes_option
@@ -12,8 +13,9 @@ from code42cli.cmds.shared import get_user_id
 from code42cli.errors import Code42CLIError
 from code42cli.file_readers import read_csv_arg
 from code42cli.file_readers import read_flat_file_arg
+from code42cli.options import format_option
 from code42cli.options import sdk_options
-
+from code42cli.output_formats import OutputFormatter
 
 DATE_FORMAT = "%Y-%m-%d"
 
@@ -23,6 +25,18 @@ DATE_FORMAT = "%Y-%m-%d"
 def departing_employee(state):
     """For adding and removing employees from the departing employees detection list."""
     pass
+
+
+@departing_employee.command("list")
+@format_option
+@sdk_options()
+def _list(state, format):
+    employee_generator = _get_departing_employees(state.sdk)
+    employee_list = [
+        employee for employees in employee_generator for employee in employees["items"]
+    ]
+    formatter = OutputFormatter(format, HEADER_KEYS_MAP)
+    formatter.echo_formatted_list(employee_list)
 
 
 @departing_employee.command()
@@ -126,6 +140,10 @@ def bulk_remove(state, file_rows):
         file_rows,
         progress_label="Removing users from departing employee detection list:",
     )
+
+
+def _get_departing_employees(sdk):
+    return sdk.detectionlists.departing_employee.get_all()
 
 
 def _add_departing_employee(sdk, username, cloud_alias, departure_date, notes):